refactor: move search query select to shared
